Update LRSDAY.14.Gene_Orthology_Identification.sh
With the installation on Proteinorto v 6.0.25, proposed on the installer a few change are necessary to meet the new functions and outputs. 
So, for example proteinortho5.pl moved to proteinortho6.pl, the option  -blastpath was replaced by -p; -project does not accept name with a . (dot) in the name and other few little changes to meet the new output naming.